Create lecture links from new schema

+++ app/database/migrations/2014_04_04_062157_create_book_abbrev_table.php
@@ -17,7 +17,7 @@ class CreateBookAbbrevTable extends Migration {
 		{
 			$table->increments('id');
             $table->string("abbrev");
-            $table->tinyInteger("bookId");
+            $table->unsignedInteger("bookId");
         });
         Log::info("Insert data...");
         $migrationsPath = app_path('database/migrations');
